refactor: myPage를 위한 로그인 유저 정보 받아오는  axios 연결

+++ backend/src/main/java/com/ssafy/api/controller/GuideController.java
@@ -15,17 +15,14 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.ssafy.api.request.BoardRequest;
-import com.ssafy.api.response.BoardListResponse;
 import com.ssafy.api.response.BoardResponse;
 import com.ssafy.api.service.BoardService;
 import com.ssafy.api.service.UserService;
 import com.ssafy.common.model.response.BaseResponseBody;
 import com.ssafy.db.entity.Board;
 import com.ssafy.db.entity.User;
-import com.ssafy.db.repository.BoardRepository;
 import com.ssafy.infos.Authority;
 import com.ssafy.infos.BoardLargeCategory;
-import com.ssafy.infos.BoardMiddleCategory;
 
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
@@ -36,77 +33,44 @@ import io.swagger.annotations.ApiResponses;
 /**
  * 유저 관련 API 요청 처리를 위한 컨트롤러 정의.
  */
-@Api(value = "커뮤니티 API", tags = { "Community" })
+@Api(value = "guide API", tags = { "Guide" })
 @RestController
-@RequestMapping("/api/community")
-public class CommunityController {
+@RequestMapping("/api/guide")
+public class GuideController {
 
 	@Autowired
 	BoardService boardService;
 
 	@Autowired
 	UserService userService;
-	
+
 	@PostMapping("/create")
 	@ApiOperation(value = "board 글등록", notes = "<strong>글 정보</strong>를 통해 게시글을 추가한다.")
 	@ApiResponses({ @ApiResponse(code = 200, message = "성공"), @ApiResponse(code = 401, message = "인증 실패"),
 			@ApiResponse(code = 404, message = "게시물 없음"), @ApiResponse(code = 500, message = "서버 오류") })
 	public ResponseEntity<? extends BaseResponseBody> createBoard(
 			@RequestBody @ApiParam(value = "글 정보", required = true) BoardRequest boardInfo) {
-		
+		boardInfo.setCategoryLarge(BoardLargeCategory.GUIDE.ordinal());
+
 		boardService.createBoard(boardInfo);
+
 		return ResponseEntity.status(200).body(BaseResponseBody.of(200, "Success"));
 	}
-	
-	@PostMapping("/reply/create")
-	@ApiOperation(value = "board 댓글 등록", notes = "<strong>글 정보</strong>를 통해 게시글을 추가한다.")
-	@ApiResponses({ @ApiResponse(code = 200, message = "성공"), @ApiResponse(code = 401, message = "인증 실패"),
-			@ApiResponse(code = 404, message = "게시물 없음"), @ApiResponse(code = 500, message = "서버 오류") })
-	public ResponseEntity<? extends BaseResponseBody> createReply(
-			@RequestBody @ApiParam(value = "글 정보", required = true) BoardRequest boardInfo) {
-		if( boardInfo.getCategoryMiddle()==BoardMiddleCategory.QnA.ordinal()) {
-			User user = userService.getUserByUid(boardInfo.getUserUid());
-			if (user.getAuthority().equals(Authority.MANAGER.toString())) {
-				boardInfo.setCategoryLarge(BoardLargeCategory.COMMUNITY.ordinal());
-				boardInfo.setCategoryMiddle(BoardMiddleCategory.REPLY.ordinal());
-				boardService.createBoard(boardInfo);
-				return ResponseEntity.status(200).body(BaseResponseBody.of(200, "Success"));
-			}
-			else return ResponseEntity.status(403).body(BaseResponseBody.of(403, "forbidden"));
-		}
-		else return ResponseEntity.status(400).body(BaseResponseBody.of(403, "bad request"));
-	}
-	
-	
-	@GetMapping("/test")
-	@ApiOperation(value = "board 검색 정보", notes = "<strong>board 를 title로 검색한 정보</strong>")
-	@ApiResponses({ @ApiResponse(code = 200, message = "성공"), @ApiResponse(code = 401, message = "인증 실패"),
-			@ApiResponse(code = 404, message = "게시물 없음"), @ApiResponse(code = 500, message = "서버 오류") })
-	public ResponseEntity<? extends BoardListResponse> test(
-			 @ApiParam(value = "board uid", required = true) @RequestParam("title") String title) {
-		
-		List<Board> boards = boardService.findBoardByTitle(title);
-		if (boards==null) {
-			return ResponseEntity.status(400).body(BoardListResponse.of(400, "Bad responce",boards));
-		}
-		else return ResponseEntity.status(200).body(BoardListResponse.of(200, "Success",boards));			
-	}
-	
+
 	@GetMapping("")
 	@ApiOperation(value = "board 검색 정보", notes = "<strong>board 를 uid로 검색한 정보</strong>")
 	@ApiResponses({ @ApiResponse(code = 200, message = "성공"), @ApiResponse(code = 401, message = "인증 실패"),
 			@ApiResponse(code = 404, message = "게시물 없음"), @ApiResponse(code = 500, message = "서버 오류") })
 	public ResponseEntity<? extends BoardResponse> findBoardByUid(
-			 @ApiParam(value = "board uid", required = true) @RequestParam("uid") int uid) {
+			@ApiParam(value = "board uid", required = true) @RequestParam("uid") int uid) {
 
 		Board board = boardService.findBoardByUid(uid);
-		if (board==null) {
-			return ResponseEntity.status(400).body(BoardResponse.of(400, "Bad responce",board));
+		if (board == null) {
+			return ResponseEntity.status(400).body(BoardResponse.of(400, "Bad responce", board));
+		} else {
+			return ResponseEntity.status(200).body(BoardResponse.of(200, "Success", board));
 		}
-		else return ResponseEntity.status(200).body(BoardResponse.of(200, "Success",board));			
 	}
-	
-	
 
 	@PutMapping("")
 	@ApiOperation(value = "board 글 변경 내용", notes = "<strong>board 를 uid로 검색한 정보</strong>.")
@@ -114,20 +78,24 @@ public class CommunityController {
 			@ApiResponse(code = 404, message = "게시물 없음"), @ApiResponse(code = 500, message = "서버 오류") })
 	public ResponseEntity<? extends BoardResponse> updateBoardByUid(
 			@RequestBody @ApiParam(value = "글 정보", required = true) BoardRequest boardInfo) {
-		Board board = boardService.findBoardByUid(boardInfo.getUid());
-		
-		if (board==null) {
-			System.out.println("업데이트 할 게시물이 없습니다.");
-			return ResponseEntity.status(400).body(BoardResponse.of(400, "not to do update",null));
-		}
-		else if(boardInfo.getUserUid()!=board.getUserUid()) 
-			return ResponseEntity.status(403).body(BoardResponse.of(403, "your authority can`t change board",board));
-		else {
-			boardService.updateBoard(board, boardInfo);
-			return ResponseEntity.status(200).body(BoardResponse.of(200, "Success",board));
-		}
+
+		System.out.println("it is in");
+		System.out.println(Authority.GENERAL.toString());
+
+		User user = userService.getUserByUid(boardInfo.getUserUid());
+		System.out.println(user.getAuthority());
+		if (user.getAuthority().equals(Authority.MANAGER.toString())) {
+			System.out.println("im in~~~~~");
+			Board board = boardService.findBoardByUid(boardInfo.getUid());
+			if (board != null) {
+				boardService.updateBoard(board, boardInfo);
+				return ResponseEntity.status(200).body(BoardResponse.of(200, "Success", board));
+			} else
+				return ResponseEntity.status(400).body(BoardResponse.of(400, "bad request", board));
+		} else
+			return ResponseEntity.status(403).body(BoardResponse.of(403, "forbidden ", null));
+
 	}
-	
 
 	@DeleteMapping("")
 	@ApiOperation(value = "board 글 삭제", notes = "<strong>board 를 uid로 검색한 게시글 삭제</strong>.")
@@ -136,19 +104,19 @@ public class CommunityController {
 	public ResponseEntity<? extends BaseResponseBody> deleteBoardByUid(
 			@ApiParam(value = "board uid", required = true) @RequestParam("uid") int uid) {
 		Board board = boardService.findBoardByUid(uid);
-		
-		if (board==null) {
-			System.out.println("삭제 할 게시물이 없습니다.");
-			return ResponseEntity.status(400).body(BaseResponseBody.of(400, "not to do update"));
-		}
-		else if(board.getUserUid()!=board.getUserUid()) 
-			return ResponseEntity.status(403).body(BaseResponseBody.of(403, "your authority can`t change board"));
-		else {
+
+		if (board == null)
+			return ResponseEntity.status(400).body(BaseResponseBody.of(400, "bad request"));
+
+		User user = userService.getUserByUid(board.getUserUid());
+
+		// board uid로 uid 찾음
+		if (user.getAuthority().equals(Authority.MANAGER.toString())) {
 			boardService.deleteBoardByUid(board);
 			return ResponseEntity.status(200).body(BaseResponseBody.of(200, "Success"));
-		}
+		} else
+			return ResponseEntity.status(403).body(BaseResponseBody.of(403, "your authority can`t delete board"));
 	}
-	
 
 	@GetMapping("/list")
 	@ApiOperation(value = "board list 정보", notes = "<strong>전체 board 목록</strong>")
@@ -156,8 +124,10 @@ public class CommunityController {
 			@ApiResponse(code = 404, message = "게시물 없음"), @ApiResponse(code = 500, message = "서버 오류") })
 	public ResponseEntity<List<Board>> getBoardList() {
 		List<Board> boards = boardService.getAllBoard();
-		if (boards==null) 
-			return  new ResponseEntity<List<Board>>(boards, HttpStatus.BAD_REQUEST);
-		else return new ResponseEntity<List<Board>>(boards, HttpStatus.OK);
+		if (boards == null) {
+			return new ResponseEntity<List<Board>>(boards, HttpStatus.BAD_REQUEST);
+		} else {
+			return new ResponseEntity<List<Board>>(boards, HttpStatus.OK);
+		}
 	}
 }
